requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
tzdata==2025.1
djangorestframework
drf-yasg
djangorestframework-simplejwt
django-allauth
dj-rest-auth
psycopg2-binary
requests
cryptography
django-cors-headers
